requirements.txt
flask
python-dotenv
openai
aiohttp
requests
pytest
pandas
flask-sqlalchemy
flask-migrate
pytest-flask